Remove other unnecessary casts to float

+++ software/o_c_REV/streams_lorenz_generator.h
@@ -55,8 +55,10 @@ class LorenzGenerator {
 
  
   inline void set_rho1(int16_t rho) {
-    rho1_ = ((double)(rho) * (1 << 13)) + 24.0 * (1 << 24) ; // was 12
-    c1_ = (double)(rho + (6 << 3)) * (1 << 13) ; // was 13
+    // rho1_ = ((double)(rho) * (1 << 13)) + 24.0 * (1 << 24) ; // was 12
+    // c1_ = (double)(rho + (6 << 3)) * (1 << 13) ; // was 13
+    rho1_ = (rho * (1 << 13)) + 24.0 * (1 << 24) ; // was 12
+    c1_ = (rho + (6 << 3)) * (1 << 13) ; // was 13
   }
 
   inline void set_rho2(int16_t rho) {
